requirements.txt
albumentations==1.4.18
numpy==1.24.4
opencv-python==4.7.0.72
matplotlib==3.7.1
scipy==1.10.1
scikit-learn==1.2.2
tqdm==4.51.0
pillow==9.5.0
timm==0.5.4
seaborn
prefetch_generator
imageio
yacs
tensorboardX
onnx==1.18.0
onnxscript==0.3.0
onnxruntime-gpu==1.22.0
onnxsim==0.4.36